func to calculate time in job based on backend data, & displayed in component
